% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_basicOF.R
\name{assert.basicOF}
\alias{assert.basicOF}
\title{test if everything is OK with the basic objective function list}
\usage{
assert.basicOF(basicOF)
}
\description{
returns error messages if the basicOF object is not a list or 
does not contain all the basic OFs or 
if it contains more then the basic OFs
}

