% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_dataframe.R
\name{assert.dataframe}
\alias{assert.dataframe}
\title{test if data is a data frame}
\usage{
assert.dataframe(data)
}
\description{
returns error message if the input: "data" is not of class "data.frame"
}

