% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_complete-Date.R
\name{channel.complete_date}
\alias{channel.complete_date}
\title{complete the date-formats with xts or COSdate}
\usage{
channel.complete_date(runoff_data)
}
\arguments{
\item{runoff_data}{The data.frame, which contains the runoff information}
}
\value{
The new runoff data.frame with the added data-format.
}
\description{
Completes the data formats of the runoff_data data.frame by adding either the needed xts-dates or COSdates
}

