% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_only-Observed.R
\name{channel.only_observed}
\alias{channel.only_observed}
\title{remove basins withouth observations}
\usage{
channel.only_observed(runoff_data)
}
\arguments{
\item{runoff_data}{A raw runoff_data data.frame, which may containts basins withouth observations.
\strong{Note:} It is assumed that all available basins are simulated!}
}
\value{
data.frame without the observation-free basins
}
\description{
Removes basins withouth observation (-999/NA values) away from the provided dataframe
}

