% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel.onlyObserved}
\alias{channel.onlyObserved}
\title{channels the basins withouth observation away}
\usage{
channel.onlyObserved(d_AllBasins)
}
\arguments{
\item{d_AllBasins}{A raw runoff dataframe, which may containts basins withouth observations. 
\strong{Note:} It is assumed that all available basins are simulated!}
}
\value{
data.frame without the observation-free basins
}
\description{
Removes basins withouth observation (-999/NA values) away from the provided dataframe
}

