% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel.path}
\alias{channel.path}
\title{Removes the filename from a path}
\usage{
channel.path(filepath)
}
\arguments{
\item{filepath}{path to a given file}
}
\value{
path to the file
}
\description{
Removes the filename from a given path
}

