% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_periods.R
\name{channel.periods}
\alias{channel.periods}
\title{calculate periods defined by user}
\usage{
channel.periods(runoff_data, start_month, end_month)
}
\arguments{
\item{runoff_data}{The data.frame, which contains the runoff information}
}
\value{
The runoff data.frame reduced and ordered according to the hydrological years within the data. 
\strong{Note:} The hydrological years are formatted as characters.
}
\description{
calculate periods defined by user
}

