% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel}
\alias{channel}
\title{Wraps the channel functions}
\usage{
channel(this, from_that)
}
\examples{
# get runoff example and clean data, remove chunk and clean names
d_raw <- fetch_runoff_example()
d_runoff <- channel(remove_chunk, runoff_data)
names(d_raw)
names(d_runoff)
}

