% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_remove-Chunk.R
\name{channel.remove_chunk}
\alias{channel.remove_chunk}
\title{removes chunk in runoff_data}
\usage{
channel.remove_chunk(runoff_data)
}
\arguments{
\item{runoff_data}{data.frame object containing at lesast COSdate, Qsim and Qobs (see: xxx)}
}
\value{
data.frame object withouth the chunk
}
\description{
Removes all collumns which are not foreseen in the runoff_data format (see: xxx)
}

