% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel.removeBloat}
\alias{channel.removeBloat}
\title{removes bloat in runoff_data}
\usage{
channel.removeBloat(runoff_data)
}
\arguments{
\item{runoff_data}{data.frame object containing at lesast COSdate, Qsim and Qobs (see: xxx)}
}
\value{
data.frame object withouth the bloat
}
\description{
Removes all collumns which are not foreseen in the runoff_data format (see: xxx)
}

