% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_evalperiods.R
\name{channel_evalPeriods}
\alias{channel_evalPeriods}
\title{calculate evaluation periods defined by user with start month and end month}
\usage{
channel_evalPeriods(runoff_data, smonth, emonth)
}
\arguments{
\item{runoff_data}{The data.frame, which contains the runoff information}
}
\value{
The runoff data.frame, including a column indicating the evaluating period 
\strong{Note:} The hydrological years are formatted as characters.
}
\description{
depreciated!
}

