% Please edit documentation in R/cook_data-GEN.R
\name{cos_data_as_xts}
\alias{cos_data_as_xts}
\title{Convert cos_data to xts-format}
\usage{
cos_data_as_xts(cos_data)
}
\arguments{
\item{cos_data}{data_frame of the cos_data (see: xxx)}
}
\value{
xts object of the cos_data data.frame
}
\description{
Converts the cos_data (class: data_frame) into an xts object
}
