% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearlyOfun-bar.R
\name{dive_expanded_barplots}
\alias{dive_expanded_barplots}
\title{hydyearly list of expaned barpltos}
\usage{
dive_expanded_barplots(Ofun_hydyearly, hydyears_in_data, num_basins, plt_ctrl)
}
\description{
creates a list of bar-plots for a the yearly objective functions (OF) xxx
see: description-follows-soonTM xxx
}

