% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearly-ofun.R
\name{dive_period_KGE}
\alias{dive_period_KGE}
\title{KGE raster plot for the available hydrological years
xxx description to follow}
\usage{
dive_period_KGE(from, given, ...)
}
\description{
KGE raster plot for the available hydrological years
xxx description to follow
}

