% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearly-ofun.R
\name{dive_period_NSE}
\alias{dive_period_NSE}
\title{NSE raster plot for the available hydrological years
xxx description to follow}
\usage{
dive_period_NSE(from, given, ...)
}
\description{
NSE raster plot for the available hydrological years
xxx description to follow
}

