% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_d-num.R
\name{fetch.d_num}
\alias{fetch.d_num}
\title{Get numbers of the different basins}
\usage{
fetch.d_num(runoff_data)
}
\arguments{
\item{runoff_data}{runoff data_frame (see: xxx)}
}
\value{
vector of integers containing the numbers of the given basins
}
\description{
Get numbers of the different basins
}

