% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_hyd-OF.R
\name{fetch.hydOF}
\alias{fetch.hydOF}
\title{Get some objective functions (OF)}
\usage{
fetch.hydOF(obs, sim)
}
\value{
data.frame contianing basic OF
}
\description{
Get some basic objective functions used in hydrology, i.e.: Root Mean Squared Error, Correlation, NSE, KGE, pbias
}

