% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_hydyears.R
\name{fetch.hydyears}
\alias{fetch.hydyears}
\title{Get hydrological years of the runoff_data}
\usage{
fetch.hydyears(runoff_data, years)
}
\value{
list with (short named) hydrological years
}
\description{
Get hydrological years of the runoff_data
}

