% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{get different data}
\usage{
fetch(what, ...)
}
\description{
more descirption goes here xxx
}
\details{
and here xxx
\itemize{
 \item{ \code{fetch(ctrl)} } { Wraps: \code{fetch.ctrl} }
 \item{ \code{fetch(control)} } {Alternvative formulation for fetch(ctrl), wraps: \code{fetch.ctrl} }
 \item{ \code{fetch(spinup,filepath, pattern)} } Wraps: \code{fetch.spinup}
 \item{ \code{fetch(number_of_basins,runoff_data)} } { Wraps: \code{fetch.number_of_basins} }
 \item{ \code{hydears,runoff_data, years_in_data} } { Wraps: \code{fetch.hydyears}} 
 \item{ \code{years_in_data, runoff_data} } { Wraps: \code{fetch.years_in_data} }
 }
}
\examples{
# get some example data
runoff_data <- fetch(runoff_example) 
head(runoff_data)
# 
runoff_data <- fetch(runoff_example) 
fetch(number_of_basins,channel.removeChunk(runoff_data))
#
runoff_data <- fetch(runoff_example) 
fetch(years_in_data,runoff_data)
}

