% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_runoff-data-RegEx.R
\name{fetch.runoff_dataRegEx}
\alias{fetch.runoff_dataRegEx}
\title{RegEx Pattern for runoff_data
 
returns the regular expression needed to check if there is Chunk in the the runoff_data data.frame(see: xxx)}
\usage{
fetch.runoff_dataRegEx()
}
\description{
RegEx Pattern for runoff_data
 
returns the regular expression needed to check if there is Chunk in the the runoff_data data.frame(see: xxx)
}

