% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_some_ofun_4_hydyears.R
\name{fetch.some_ofun_4_hydyears}
\alias{fetch.some_ofun_4_hydyears}
\title{Get basic objective function for runoff_data}
\usage{
fetch.some_ofun_4_hydyears(runoff_data, hydyears_in_data)
}
\arguments{
\item{runoff_data}{runoff_data data.frame (see:xxx).}
}
\value{
list of baisc objective function evaluated for the different hydrological years and over the whole timespan.
}
\description{
Calculate basic objective functions (i.e. NSE, KGE, percentage BIAS, COrrelation (see: xxx)) for
every basin and hydrological year
}

