% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_years-in-data.R
\name{fetch.yearsindata}
\alias{fetch.yearsindata}
\title{Get years of the runoff_data}
\usage{
fetch.yearsindata(runoff_data)
}
\value{
list with (short and long named) years. \strong{Note:} Short names years are just characters.
Thus, they may only be usefull for plotting
}
\description{
Get years of the runoff_data
}

