% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_hydyears.R
\name{fetch_hydyears}
\alias{fetch_hydyears}
\title{Get hydrological years of the runoff_data}
\usage{
fetch_hydyears(runoff_data, years_in_data)
}
\arguments{
\item{runoff_data}{runoff data.frame as defined in xxx}

\item{years_in_data}{list as returned by \code{\link[visCOS]{fetch_years_in_data}}}
}
\value{
list with (short named) hydrological years
}
\description{
Get hydrological years of the runoff_data
}

