% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_number_of_basins.R
\name{fetch_number_of_basins}
\alias{fetch_number_of_basins}
\title{Get numbers of the different basins}
\usage{
fetch_number_of_basins(runoff_data)
}
\arguments{
\item{runoff_data}{runoff data_frame (see: xxx)}
}
\value{
vector of integers containing the numbers of the given basins
}
\description{
Get numbers of the different basins
}

