% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_regex_for_runoff_data.R
\name{fetch_regex_for_runoff_data}
\alias{fetch_regex_for_runoff_data}
\title{RegEx Pattern for runoff_data}
\usage{
fetch_regex_for_runoff_data()
}
\description{
returns the regular expression needed to check if there is Chunk in the the runoff_data data.frame(see: xxx)
}

