% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_runoff_as_xts.R
\name{fetch_runoff_as_xts}
\alias{fetch_runoff_as_xts}
\title{Convert runoff_data to xts-format}
\usage{
fetch_runoff_as_xts(runoff_data)
}
\arguments{
\item{runoff_data}{data_frame of the runoff_data (see: xxx)}
}
\value{
xts object of the runoff_data data.frame
}
\description{
Converts the runoff_data (class: data_frame) into an xts object
}

