% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implode_Cosdate.R
\name{implode.Cosdate}
\alias{implode.Cosdate}
\title{transforms COSdate into the nicer POSIXct-date format}
\usage{
implode.Cosdate(data_frame)
}
\arguments{
\item{data_frame}{A data.frame object containing a time series of the COSdate format}
}
\value{
dates in the format of the POSIXct-class
}
\description{
Takes a data.frame, which contains the COSdate format (see: xxxx) and transforms it into a POSIXct series.
Note that time is assumbed to be in UTC
}

