% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_expanded_barplots.R, R/pour.R
\name{list_yOF_barplts}
\alias{list_yOF_barplts}
\title{creates a list of bar-plots for a the yearly objective functions (OF)
see: description-follows-soonTM
***}
\usage{
list_yOF_barplts(OF_hydyearly, eval_size, d_nums, plt_ctrl)

list_yOF_barplts(OF_hydyearly, eval_size, d_nums, plt_ctrl)
}
\description{
creates a list of bar-plots for a the yearly objective functions (OF)
see: description-follows-soonTM
***

creates a list of bar-plots for a the yearly objective functions (OF)
see: description-follows-soonTM
***
}

