% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof-GEN.R
\name{mof_barplot}
\alias{mof_barplot}
\alias{mof_rasterplot}
\title{Bar plot for the Main Objective Function Values}
\usage{
mof_barplot(cos_data)

mof_rasterplot(cos_data)
}
\description{
Bar plot for the Main Objective Function Values

Bar plot for the Main Objective Function Values
}
