% Please edit documentation in R/of_window-GEN.R
\name{of_windows}
\alias{of_windows}
\title{Objectie Function Windowing}
\usage{
of_windows(cos_data, of, lb = 0, min_window_size = 500, max_window_size = 1000, 
    number_windows = 3, na_filling = FALSE)
}
\description{
Compute the objective function for a set of windows.
}
