% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot-GEN.R, R/plotlist_runoffpeaks-GEN.R
\name{peak_plot}
\alias{peak_plot}
\title{Plot List for Runoff Peaks}
\usage{
peak_plot(runoff_data, n_events = 10L, window_size = 24L)

peak_plot(runoff_data, n_events = 10L, window_size = 24L)
}
\description{
Plot List for Runoff Peaks

Plot List for Runoff Peaks
}

