% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{periodplot_CORR}
\alias{periodplot_CORR}
\title{Correlation raster plot for the available hydrological years
xxx description to follow}
\usage{
periodplot_CORR(runoff_data, plt_ctrl)
}
\description{
Correlation raster plot for the available hydrological years
xxx description to follow
}

