% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{periodplot_KGE}
\alias{periodplot_KGE}
\title{KGE raster plot for the available hydrological years
xxx description to follow}
\usage{
periodplot_KGE(runoff_data, plot_title = "Yearly KGE",
  y_label = "basin number")
}
\description{
KGE raster plot for the available hydrological years
xxx description to follow
}

