% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{periodplot_NSE}
\alias{periodplot_NSE}
\title{NSE raster plot for the available hydrological years
xxx description to follow}
\usage{
periodplot_NSE(runoff_data, plot_title = "yearly NSE",
  y_label = "basin number")
}
\description{
NSE raster plot for the available hydrological years
xxx description to follow
}

