% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{periodplot_pBIAS}
\alias{periodplot_pBIAS}
\title{Percentage Bias raster plot for the available hydrological years
xxx description to follow}
\usage{
periodplot_pBIAS(runoff_data, plt_ctrl)
}
\description{
Percentage Bias raster plot for the available hydrological years
xxx description to follow
}

