% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_of-GEN.R, R/main_of-GEN.R
\name{plot_main_of}
\alias{plot_main_of}
\title{Plot main objective fucnction values}
\description{
Currently two options for plotting the main objectives are provided by 
visCOS: Plotting the differnet objective functions values as a set of 
barplots \code{barplot_of} and plotting a summary table in form of 
a large raster of all the objective fucntion values \code{rasterplot_of}.

Currently two options for plotting the main objectives are provided by 
visCOS: Plotting the differnet objective functions values as a set of 
barplots \code{barplot_of} and plotting a summary table in form of 
a large raster of all the objective fucntion values \code{rasterplot_of}.
}

