% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlist_runoffpeaks-GEN.R
\name{plotlist_runoffpeaks}
\alias{plotlist_runoffpeaks}
\title{Plot List for Runoff Peaks}
\usage{
plotlist_runoffpeaks(runoff_data, n_events = 10L, window_size = 24L)
}
\description{
Plot List for Runoff Peaks
}

