% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_save-expanded-barplots.R
\name{pour.expandedbars.intoFile}
\alias{pour.expandedbars.intoFile}
\title{save the the list barplots}
\usage{
pour.expandedbars.intoFile(list_of_barplts, path = "",
  jpg_filenames = "bar_plot", hmtl_filename = "summary_file")
}
\description{
saves the list returned by pour.expanded_barplots xxx more description will follow TM
}

