% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearly-ofun.R
\name{pour.hydyearly_KGE}
\alias{pour.hydyearly_KGE}
\title{KGE raster plot for the available hydrological years
xxx description to follow}
\usage{
pour.hydyearly_KGE(from, given, ...)
}
\description{
KGE raster plot for the available hydrological years
xxx description to follow
}

