% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearlyOfun-bar.R
\name{pour.list_of_ofun_barplots}
\alias{pour.list_of_ofun_barplots}
\title{yearly runoff list}
\usage{
pour.list_of_ofun_barplots(Ofun_hydyearly, hydyears_in_data, num_basins,
  plt_ctrl)
}
\description{
creates a list of bar-plots for a the yearly objective functions (OF) xxx
see: description-follows-soonTM xxx
}

