% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{pour}
\alias{pour}
\title{get different data}
\usage{
pour(what, ...)
}
\description{
The \code{pour} function is a wrapper around the pour_ functions. It works as following: 
\itemize{
 \item{ \code{pour(spinup,filepath, pattern)} } { Wraps: \code{\link[visCOS]{pour_spinup}} }
 \item{ \code{pour(number_of_basins,runoff_data)} } { Wraps: \code{\link[visCOS]{pour_number_of_basins}} }
 \item{ \code{pour(hydears,runoff_data, years_in_data)} } { Wraps: \code{\link[visCOS]{pour_hydyears}} }
 \item{ \code{pour(years_in_data, runoff_data)} } { Wraps: \code{\link[visCOS]{pour_years_in_data}} }
 \item{ \code{pour(runoff_as_xts, runoff_data)} } { Wraps: \code{\link[visCOS]{pour_runoff_as_xts}} }
 }
}
\examples{
# get some example data:
runoff_data <- pour(runoff_example) 
head(runoff_data)
# 
# get number of basins: 
runoff_data <- pour(runoff_example) 
pour(number_of_basins,prepare_removeChunk(runoff_data))
#
# get years in data 
runoff_data <- pour(runoff_example) 
pour(years_in_data,runoff_data)
}

