% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearly-ofun.R
\name{pour.yearly_ofun}
\alias{pour.yearly_ofun}
\title{ggplot wrapper for the hydyearly objective functions}
\usage{
pour.yearly_ofun(bOF, choice = "NSE", hydyears_in_d, plt_ctrl)
}
\arguments{
\item{bOF}{list, as returned by \code{\link[visCOS]{fetch.basicOfun}}}

\item{hydyears_in_d}{hydrears in data, as returned by \code{\link[visCOS]{fetch.hydyears}}}

\item{string}{with the chosen baisc objective function.
\code{\link[visCOS]{fetch.basicOfun}} provides "NSE", "KGE", "pBIAS" or "CORR"}

\item{xxx}{yet to be defined control list}
}
\description{
plot table of the yearly basic objective function
}

