% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearlyOfun-bar.R
\name{pour.yearlyOfun_bar}
\alias{pour.yearlyOfun_bar}
\title{yearly runoff list}
\usage{
pour.yearlyOfun_bar(Ofun_hydyearly, eval_size, d_nums, plt_ctrl)
}
\description{
creates a list of bar-plots for a the yearly objective functions (OF)
see: description-follows-soonTM
***
}

