% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_yearly-ofun.R
\name{pour_period_pBIAS}
\alias{pour_period_pBIAS}
\title{Percentage Bias raster plot for the available hydrological years
xxx description to follow}
\usage{
pour_period_pBIAS(from, given, ...)
}
\description{
Percentage Bias raster plot for the available hydrological years
xxx description to follow
}

