% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.plt_ctrl.R
\name{pour_plt_ctrl}
\alias{pour_plt_ctrl}
\title{get basic config for the plot control}
\usage{
pour_plt_ctrl(plot_title = "Title", ltitle = "Legend", year_name = "Year",
  colors = c("#FF3300", "#f6f3a1", "#005900", "purple4"),
  color_midpoint = 0.5, limits = c(0, 1), lb_cut = 0, text_size = 0.5)
}
\description{
xxx description has to follow
}

