% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_regex_for_runoff_data.R
\name{pour_regex_for_runoff_data}
\alias{pour_regex_for_runoff_data}
\title{RegEx Pattern for runoff_data}
\usage{
pour_regex_for_runoff_data()
}
\description{
returns the regular expression needed to check if there is Chunk in the the runoff_data data.frame(see: xxx)
}

