% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_some_ofun.R
\name{pour_some_ofun}
\alias{pour_some_ofun}
\title{Get some objective functions (OF)}
\usage{
pour_some_ofun(obs, sim)
}
\value{
data.frame contianing basic OF
}
\description{
Get some basic objective functions used in hydrology, i.e.: Root Mean Squared Error, Correlation, NSE, KGE, pbias
}

