% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_spinup.R
\name{pour_spinup}
\alias{pour_spinup}
\title{Extract spinup time from stats file}
\usage{
pour_spinup(filepath, pattern)
}
\arguments{
\item{filepath}{path to the given file}

\item{pattern}{string used to mark that the next integer is the spinup time}
}
\value{
Integer indicating the length of the spin-up time in hours
}
\description{
A little wrapper to get information about the spinup position from a file. The spinup-time is supposed to be an integer, which marks the position in the data which can be cut away for various reasons (e.g. model spin up time).
The wrapper assumes that the infomration (integer) is stored in a file (filepath) and marked by pattern, e.g.: within the string "Pre-Amp-Time: 200" the frist part would be the `pattern`.
}

