% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_periods.R
\name{prepare.periods}
\alias{prepare.periods}
\title{calculate periods}
\usage{
prepare.periods(runoff_data, start_month, end_month)
}
\arguments{
\item{runoff_data}{The data.frame, which contains the runoff information}
}
\value{
The runoff data.frame reduced and ordered according to the hydrological years within the data. 
\strong{Note:} The hydrological years are formatted as characters.
}
\description{
Mark the periods within runoff_data. The range of the period is defined monthly by the integers start_month and end_month.
}

