% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hydyears.R
\name{prepare_hydyears}
\alias{prepare_hydyears}
\title{calculate hydrological years}
\usage{
prepare_hydyears(runoff_data)
}
\arguments{
\item{runoff_data}{The data.frame, which contains the runoff information}
}
\value{
The runoff data.frame reduced and ordered according to the hydrological years within the data. 
\strong{Note:} The hydrological years are formatted as characters.
}
\description{
calculate hydrological years
}

