% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_implode-Cosdate.R
\name{prepare_implode_cosdate}
\alias{prepare_implode_cosdate}
\title{transforms COSdate into the nicer POSIXct-date format}
\usage{
prepare_implode_cosdate(data_frame)
}
\arguments{
\item{data_frame}{A data.frame object containing a time series of the COSdate format}
}
\value{
dates in the format of the POSIXct-class
}
\description{
Takes a data.frame, which contains the COSdate format (see: xxx) and transforms it into a POSIXct series.
Note that time is assumbed to be in UTC
}
\examples{
Get runof example and implode cos_date to obtain the POSIXct date
require(magrittr)
runoff <- pour_runoff_example()
runoff \%<>\% prepare_implode_cosdate 
# not a fan of magrittr-piping? Just use: runoff <- prepare_implode_cosdate(runoff)
# 
# note that if prepare_implode_cosdate is not used on the runoff data istelf it will work like a pour and just return the posix_dates 
runoff <- pour_runoff_example()
posix_dates <- prepare_implode_cosdate(runoff)
head(posix_dates)
}

