% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_path.R
\name{prepare_path}
\alias{prepare_path}
\title{Removes the filename from a path}
\usage{
prepare_path(filepath)
}
\arguments{
\item{filepath}{path to a given file}
}
\value{
path to the file
}
\description{
Removes the filename from a given path
}

