% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearlyOfun-bar.R
\name{serve.expanded_barplots}
\alias{serve.expanded_barplots}
\title{hydyearly list of expaned barpltos}
\usage{
serve.expanded_barplots(Ofun_hydyearly, hydyears_in_data, num_basins, plt_ctrl)
}
\description{
creates a list of bar-plots for a the yearly objective functions (OF) xxx
see: description-follows-soonTM xxx
}

