% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_ofun.R
\name{serve.ofun}
\alias{serve.ofun}
\title{Get some objective functions (OF)}
\usage{
serve.ofun(obs, sim)
}
\value{
data.frame contianing basic OF
}
\description{
Get some basic objective functions used in hydrology, i.e.: Root Mean Squared Error, Correlation, NSE, KGE, pbias
}

