% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve.period_NSE}
\alias{serve.period_NSE}
\title{NSE raster plot for the available hydrological years
xxx description to follow}
\usage{
serve.period_NSE(from, given, ...)
}
\description{
NSE raster plot for the available hydrological years
xxx description to follow
}

