% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve.period_pBIAS}
\alias{serve.period_pBIAS}
\title{Percentage Bias raster plot for the available hydrological years
xxx description to follow}
\usage{
serve.period_pBIAS(from, given, ...)
}
\description{
Percentage Bias raster plot for the available hydrological years
xxx description to follow
}

