% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve.plot_period_CORR}
\alias{serve.plot_period_CORR}
\title{Correlation raster plot for the available hydrological years
xxx description to follow}
\usage{
serve.plot_period_CORR(runoff_data, plt_ctrl)
}
\description{
Correlation raster plot for the available hydrological years
xxx description to follow
}

