% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve.plot_period_KGE}
\alias{serve.plot_period_KGE}
\title{KGE raster plot for the available hydrological years
xxx description to follow}
\usage{
serve.plot_period_KGE(runoff_data, plt_ctrl)
}
\description{
KGE raster plot for the available hydrological years
xxx description to follow
}

