% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve.plot_period_NSE}
\alias{serve.plot_period_NSE}
\title{NSE raster plot for the available hydrological years
xxx description to follow}
\usage{
serve.plot_period_NSE(runoff_data, plt_ctrl)
}
\description{
NSE raster plot for the available hydrological years
xxx description to follow
}

