% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve.plot_period_pBIAS}
\alias{serve.plot_period_pBIAS}
\title{Percentage Bias raster plot for the available hydrological years
xxx description to follow}
\usage{
serve.plot_period_pBIAS(runoff_data, plt_ctrl)
}
\description{
Percentage Bias raster plot for the available hydrological years
xxx description to follow
}

