% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.waterbalance.R
\name{serve.plot_waterbalance}
\alias{serve.plot_waterbalance}
\title{water bilance plot}
\usage{
serve.plot_waterbalance(runoff_data, obs_name, cum_area_in_km = NULL,
  calculate_mm = TRUE)
}
\description{
pltos the water bilance of a chosen basin
}

