% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.plotlist_periodOF.R
\name{serve.plotlist_periodOF}
\alias{serve.plotlist_periodOF}
\title{hydyearly list of expaned barpltos}
\usage{
serve.plotlist_periodOF(runoff_data, plt_ctrl)
}
\description{
creates a list of bar-plots for a the yearly objective functions (OF) xxx
see: description-follows-soonTM xxx
}

