% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.waterbilance.R
\name{serve.plotlist_waterbalance}
\alias{serve.plotlist_waterbalance}
\title{create list of water-bilance plots}
\usage{
serve.plotlist_waterbalance(runoff_data, cum_area_in_km = NULL,
  calculate_mm = TRUE)
}
\description{
xxx
}

