% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.save_list3x3.R
\name{serve.save_list3x3}
\alias{serve.save_list3x3}
\title{save the the list barplots}
\usage{
serve.save_list3x3(plot_list, path = "", jpg_filenames = "bar_plot",
  hmtl_filename = "summary_file")
}
\description{
saves the list returned by serve.expanded_barplots xxx more description will follow TM
}

