% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.save_plotlist.R
\name{serve.save_plotlist}
\alias{serve.save_plotlist}
\title{save a list of barplots}
\usage{
serve.save_plotlist(plot_list, path = "", jpg_filenames = "plot",
  hmtl_filename = "summary_file")
}
\description{
saves the list of plots as .jpeg and embeds them into an .html file
}

