% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.save_list3x3.R
\name{serve.save_plotlist3x3}
\alias{serve.save_plotlist3x3}
\title{save a list of barplots}
\usage{
serve.save_plotlist3x3(plot_list, path = "", jpg_filenames = "plot",
  hmtl_filename = "summary_file")
}
\description{
#' saves the list of plots within a 3x3 raster as .jpeg and embeds them into an .html file
}

