% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_total-Ofun.R
\name{serve.totalOfun}
\alias{serve.totalOfun}
\title{define rasterplot functions for total OF}
\usage{
serve.totalOfun(choice, bOF, plt_ctrl)
}
\description{
define rasterplot functions for total OF
}

