% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.waterbalance.R
\name{serve.waterbalance}
\alias{serve.waterbalance}
\title{makes plot for the water bilance}
\usage{
serve.waterbalance(runoff_data, obs_name, cum_area_in_km = NULL,
  calculate_mm = TRUE)
}
\description{
xxx
}

