% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.waterbilance.R
\name{serve.waterbilance}
\alias{serve.waterbilance}
\title{makes plot for the water bilance}
\usage{
serve.waterbilance(runoff_data, obs_name, cum_area_in_km = NULL,
  calculate_mm = TRUE)
}
\description{
xxx
}

