% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_save-expanded-barplots.R
\name{serve_expandedbars_intoFile}
\alias{serve_expandedbars_intoFile}
\title{save the the list barplots}
\usage{
serve_expandedbars_intoFile(list_of_barplts, path = "",
  jpg_filenames = "bar_plot", hmtl_filename = "summary_file")
}
\description{
saves the list returned by serve_expanded_barplots xxx more description will follow TM
}

