% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve_period_KGE}
\alias{serve_period_KGE}
\title{KGE raster plot for the available hydrological years
xxx description to follow}
\usage{
serve_period_KGE(from, given, ...)
}
\description{
KGE raster plot for the available hydrological years
xxx description to follow
}

