% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_period_ofun.R
\name{serve_period_ofun}
\alias{serve_period_ofun}
\title{Get basic objective function for runoff_data}
\usage{
serve_period_ofun(runoff_data)
}
\arguments{
\item{runoff_data}{runoff_data data.frame (see:xxx).}
}
\value{
list of baisc objective function evaluated for the different hydrological years and over the whole timespan.
}
\description{
Calculate basic objective functions (i.e. NSE, KGE, percentage BIAS, COrrelation (see: xxx)) for
every basin and the chosen periods
}

