% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_yearly-ofun.R
\name{serve_yearly_ofun}
\alias{serve_yearly_ofun}
\title{ggplot wrapper for the hydyearly objective functions}
\usage{
serve_yearly_ofun(choice, bOF, periods_in_data, plt_ctrl)
}
\arguments{
\item{bOF}{list, as returned by \code{\link[visCOS]{pour_basicOfun}}}

\item{periods_in_data}{periods in data, as returned by \code{\link[visCOS]{pour_periods}}}

\item{string}{with the chosen baisc objective function.
\code{\link[visCOS]{pour_basicOfun}} provides "NSE", "KGE", "pBIAS" or "CORR"}

\item{xxx}{yet to be defined control list}
}
\description{
plot table of the yearly basic objective function
}

