% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{sweep.path}
\alias{sweep.path}
\title{sweeps the filename away}
\usage{
sweep.path(filepath)
}
\arguments{
\item{filepath}{path to a given file}
}
\value{
path to the file
}
\description{
Removes the filename from a path
}

