% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defensiveCode.R
\name{testfor.basicOF}
\alias{testfor.basicOF}
\title{test if everything is OK with the basic objective function list}
\usage{
testfor.basicOF(basicOF)
}
\description{
returns error messages if the basicOF object is not a list or 
does not contain all the basic OFs or 
if it contains more then the basic OFs
}

