% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defensiveCode.R
\name{testfor.Chunk}
\alias{testfor.Chunk}
\title{test for Chunk 
 
tests if there is still chunk in the runoff_data data frame}
\usage{
testfor.Chunk(runoff_data)
}
\description{
test for Chunk 
 
tests if there is still chunk in the runoff_data data frame
}

