% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defensiveCode.R
\name{testfor.dataframe}
\alias{testfor.dataframe}
\title{test if data is a data frame}
\usage{
testfor.dataframe(data)
}
\description{
returns error message if the input: "data" is not of class "data.frame"
}

