% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_total-Ofun.R
\name{totalplot_KGE}
\alias{totalplot_KGE}
\title{rasterpolots for the KGE of the whole period}
\usage{
totalplot_KGE(runoff_data, plot_titel = "Total KGE", legend_title = "KGE")
}
\description{
xxx description follows
}

