% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_total-Ofun.R
\name{totalplot_NSE}
\alias{totalplot_NSE}
\title{rasterpolots for the NSE of the whole period}
\usage{
totalplot_NSE(runoff_data, plot_title = "Total NSE", legend_title = "NSE")
}
\description{
xxx description follows
}

