\name{visCOS-imports}
\alias{.__T__$:base}
\alias{.__T__$<-:base}
\alias{.__T__[:base}
\alias{.__T__[[<-:base}
\alias{.__T__[<-:base}
\alias{\%>\%}
\docType{import}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages. Follow the links to their documentation.
\describe{
  \item{dplyr}{\code{\link[dplyr:.__T__$:base]{.__T__$:base}}, \code{\link[dplyr:.__T__$<-:base]{.__T__$<-:base}}, \code{\link[dplyr:.__T__[:base]{.__T__[:base}}, \code{\link[dplyr:.__T__[[<-:base]{.__T__[[<-:base}}, \code{\link[dplyr:.__T__[<-:base]{.__T__[<-:base}}}
  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}}
