% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/of_window-GEN.R
\name{window_of}
\alias{window_of}
\title{Objectie Function Windowing}
\usage{
window_of(cos_data, of, lb = 0, min_window_size = 500, max_window_size = 1000, 
    number_windows = 3, na_filling = FALSE)
}
\description{
Compute the objective function for a set of windows.
}
