% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive.R
\name{dive_runoff_with_ofun}
\alias{dive_runoff_with_ofun}
\title{dive with ofun 
Runs a Shiny App which can be used to get an overview of a runoff_data time series object. 
Explore the runoff_data with a little [shiny](http://shiny.rstudio.com/) App.}
\usage{
dive_runoff_with_ofun(runoff_data)
}
\arguments{
\item{d_xts}{runoff_data formatted as time series}
}
\description{
dive with ofun 
Runs a Shiny App which can be used to get an overview of a runoff_data time series object. 
Explore the runoff_data with a little [shiny](http://shiny.rstudio.com/) App.
}

