% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive.R, R/wrap_explore_runoff.R
\name{explore.runoff}
\alias{explore.runoff}
\title{Explore runoff_data time series with shiny}
\usage{
explore.runoff(d_xts)

explore.runoff(d_xts)
}
\arguments{
\item{d_xts}{runoff_data formatted as time series}

\item{d_xts}{runoff_data formatted as time series}
}
\description{
Runs a Shiny App which can be used to get an overview of a runoff_data time series object

Runs a Shiny App which can be used to get an overview of a runoff_data time series object
}

