% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{get different data}
\usage{
fetch(what, ...)
}
\description{
The \code{fetch} function is a wrapper around the fetch_ functions. It works as following: 
\itemize{
 \item{ \code{fetch(spinup,filepath, pattern)} } { Wraps: \code{\link[visCOS]{fetch_spinup}} }
 \item{ \code{fetch(number_of_basins,runoff_data)} } { Wraps: \code{\link[visCOS]{fetch_number_of_basins}} }
 \item{ \code{fetch(hydears,runoff_data, years_in_data)} } { Wraps: \code{\link[visCOS]{fetch_hydyears}} }
 \item{ \code{fetch(years_in_data, runoff_data)} } { Wraps: \code{\link[visCOS]{fetch_years_in_data}} }
 \item{ \code{fetch(runoff_as_xts, runoff_data)} } { Wraps: \code{\link[visCOS]{fetch_runoff_as_xts}} }
 }
}
\examples{
# get some example data:
runoff_data <- fetch(runoff_example) 
head(runoff_data)
# 
# get number of basins: 
runoff_data <- fetch(runoff_example) 
fetch(number_of_basins,channel_removeChunk(runoff_data))
#
# get years in data 
runoff_data <- fetch(runoff_example) 
fetch(years_in_data,runoff_data)
}

