% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{get different data}
\usage{
fetch(what, ...)
}
\description{
more descirption goes here xxx
}
\details{
and here xxx
\itemize{
 \item{ \code{fetch(ctrl)} } { wraps: \code{fetch.ctrl} }
 \item{ \code{fetch(control)} } {alternvative formulation for fetch(ctrl), wraps \code{fetch.ctrl} }
 \item{}
 \item{ \code{fetch(number_of_basins,runoff_data)} } {  }
 - fetch(spinup, filepath, pattern),wraps: fetch.spinup
 - , fetch.number_of_basins
 - fetch(hydears, runoff_data, years_in_data), wraps: fetch.hydyears
 - fetch(years_in_data, runoff_data), wraps fetch.years_in_data
 }
}
\examples{
# get some example data
runoff_data <- fetch(runoff_example) 
head(runoff_data)
# 
runoff_data <- fetch(runoff_example) 
fetch(number_of_basins,channel.removeChunk(runoff_data))
#
runoff_data <- fetch(runoff_example) 
fetch(years_in_data,runoff_data)
}

