% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{get different data}
\usage{
fetch(what, ...)
}
\description{
get different data
}
\examples{
# get some example data
runoff_data <- fetch(runoff_example) 
head(runoff_data)
# 
runoff_data <- fetch(runoff_example) 
fetch(number_of_basins,channel.removeChunk(runoff_data))
}

