% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{implode.Cosdate}
\alias{implode.Cosdate}
\title{transforms COSdate into a nice date format}
\usage{
implode.Cosdate(data_frame)
}
\arguments{
\item{data_frame}{A data.frame object containing a time series of the COSdate format}
}
\value{
dates in the format of the POSIXct-class
}
\description{
Takes a data_frame containing the COSdate format and transforms it into a POSIX date series
}

