% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data-GEN.R
\name{mark_periods}
\alias{mark_periods}
\title{calculate periods}
\usage{
mark_periods(cos_data, start_month = 10, end_month = 9)
}
\arguments{
\item{cos_data}{The data.frame, which contains the runoff information}
}
\value{
The runoff data.frame reduced and ordered according to the
hydrological years within the data.
\strong{Note:} The periods columns are formatted as characters!
}
\description{
Mark the periods within cos_data.
`start_month` and `end_month`.
}

