% Please edit documentation in R/of_compare-GEN.R
\name{of_compare}
\alias{of_compare}
\title{Explore with Objective Functions}
\usage{
of_compare(d1, d2 = NULL, of_list = list(nse = d_nse, kge = d_kge, p_bias = d_pbias, 
    r = d_cor))
}
\description{
Runs a Shiny Gadget which can be used to interactively explore two 
`cos_data` objects in terms of their respective objective functions.
}
