% Please edit documentation in R/of-GEN.R
\name{of_barplot}
\alias{of_barplot}
\alias{of_rasterplot}
\title{Bar plot for the Main Objective Function Values}
\usage{
of_barplot(cos_data, d_metrics = list(nse = d_nse, kge = d_kge, pbias = d_pbias, 
    corr = d_cor))

of_rasterplot(cos_data, d_metrics = list(nse = d_nse, kge = d_kge, pbias = d_pbias, 
    corr = d_cor))
}
\description{
Bar plot for the Main Objective Function Values

Bar plot for the Main Objective Function Values
}
