% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{pour}
\alias{pour}
\title{get different data}
\usage{
pour(what, ...)
}
\description{
The \code{pour} function is a wrapper around the pour_ functions. It works as following: 
\itemize{
 \item{ \code{pour(runoff_example)} } { Wraps: \code{\link[visCOS]{pour_runoff_example}} }
 \item{ \code{pour(spinup, filepath, pattern)} } { Wraps: \code{\link[visCOS]{pour_spinup}} }
 \item{ \code{pour(plt_ctrl)} } { Wraps: \code{\link[visCOS]{pour_plt_ctrl}} }
 }
}
\examples{
# get some example data:
runoff_data <- pour(runoff_example) 
head(runoff_data)
# get regular expression for the runoff_data 
reg_expr <- pour(regex_runoff_data)
}

