% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_spinup.R
\name{pour_spinup}
\alias{pour_spinup}
\title{Extract spinup time from stats file}
\usage{
pour_spinup(filepath, pattern)
}
\arguments{
\item{filepath}{path to the given file}

\item{pattern}{string used to mark that the next integer is the spinup time}
}
\value{
Integer indicating the length of the spin-up time in hours
}
\description{
Spinup-time is an integer. It is the time in hours, which is not used for the model-evalution. 
The pattern is the used marker within the file
}

