% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Wraps the prepare functions}
\usage{
prepare(this, from_that)
}
\examples{
# get runoff example and clean data, remove chunk and clean names
d_raw <- pour.runoff_example()
d_runoff <- prepare(remove_chunk, runoff_data)
names(d_raw)
names(d_runoff)
}

