% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_complete-Date.R
\name{prepare_complete_date}
\alias{prepare_complete_date}
\title{complete the date-formats with xts or COSdate}
\usage{
prepare_complete_date(runoff_data)
}
\arguments{
\item{runoff_data}{The data.frame, which contains the runoff information}
}
\value{
The new runoff data.frame with the added data-format.
}
\description{
complete the date-formats with xts or COSdate
}

