% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.plot_waterbilance.R
\name{serve.plot_waterbilance}
\alias{serve.plot_waterbilance}
\title{water bilance plot}
\usage{
serve.plot_waterbilance(runoff_data, obs_name, calculate_mm = TRUE,
  cum_area_in_km = NULL)
}
\description{
pltos the water bilance of a chosen basin
}

