% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{sweep.NoSim}
\alias{sweep.NoSim}
\title{sweeps the basins withouth observation away}
\usage{
sweep.NoSim(d_AllBasins)
}
\arguments{
\item{d_AllBasins}{A raw runoff dataframe, which may containts basins withouth observations. 
\strong{Note:} It is assumed that all available basins are simulated!}
}
\value{
data.frame without the observation-free basins
}
\description{
Removes basins withouth observation (-999/NA values) away from the provided dataframe
}

