% Please edit documentation in R/viscos_options-GEN.R
\docType{data}
\name{viscos_options}
\alias{viscos_options}
\title{visCOS global options}
\format{An object of class \code{GlobalOptionsFun} of length 1.}
\usage{
viscos_options
}
\description{
get and set the global options of visCOS
}
\keyword{datasets}
