% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour_plt_ctrl.R
\docType{data}
\name{viscos_options}
\alias{viscos_options}
\title{get basic config for the plot control}
\format{An object of class \code{GlobalOptionsFun} of length 1.}
\usage{
viscos_options
}
\description{
Many of the serve function of visCOS are controlled via a list containing the different plotting controls. This function provides such a list with the basic settings added.
}
\keyword{datasets}

